parse all EPO XML files tp structured json
